requirements.txt
aioredis==1.2.0
asgi-redis==1.4.3
asgiref==3.1.4
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.7.1
Automat==0.7.0
autopep8==1.4.3
bokeh==1.0.4
certifi==2019.6.16
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
Click==7.0
colorama==0.4.1
constantly==15.1.0
cryptography==2.7
cycler==0.10.0
daphne==2.3.0
dj-database-url==0.5.0
Django==2.2.3
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-extensions==2.1.9
django-postgres-copy==2.4.1
djangorestframework==3.9.4
Flask==1.1.1
gunicorn==19.9.0
hiredis==1.0.0
httpie==1.0.2
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.16
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mpld3==0.3
msgpack==0.6.1
msgpack-python==0.5.6
numpy==1.16.2
oauthlib==3.0.2
packaging==19.0
pandas==0.24.2
Pillow==5.4.1
psycopg2==2.8.3
pycodestyle==2.5.0
pycparser==2.19
pydotplus==2.0.2
Pygments==2.4.2
PyHamcrest==1.9.0
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
redis==2.10.6
requests==2.22.0
requests-oauthlib==1.2.0
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
tornado==6.0.2
Twisted==19.2.1
txaio==18.8.1
typed-ast==1.3.1
urllib3==1.25.3
Werkzeug==0.15.5
whitenoise==4.1.3
wrapt==1.11.1
zope.interface==4.6.0
